requirements.txt
fastapi
uvicorn
crawl4ai
python-dotenv
pydantic
requests















